requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
django-filter==2.4.0
psycopg2-binary==2.8.6
djoser
djangorestframework_simplejwt